sandbox/pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "black>=25.1.0",
    "boto3>=1.40.4",
    "fastapi>=0.116.1",
    "flake8>=7.3.0",
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.35.0",
]
